pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "recombinase"
version = "0.1.5"
description = "Template-guided document synthesis: extract structured content from source pptx files and recombine into a canonical template"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Terry Li" },
]
keywords = ["pptx", "powerpoint", "template", "cv", "credentials", "extraction", "generation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business",
    "Topic :: Text Processing",
]
dependencies = [
    "python-pptx>=0.6.23",
    "pyyaml>=6.0",
    "typer>=0.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.6",
    "mypy>=1.8",
    "types-PyYAML",
]

[project.scripts]
recombinase = "recombinase.cli:main"

[project.urls]
Homepage = "https://github.com/terry-li-hm/recombinase"
Issues = "https://github.com/terry-li-hm/recombinase/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/recombinase"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific
    "N",    # pep8-naming
    "PTH",  # use pathlib
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = [
    "N802",  # allow lowercase test function names
    "N806",  # allow camelCase locals that mirror OOXML element names
]
"src/recombinase/cli.py" = [
    "B008",    # Typer requires function calls in arg defaults
    "BLE001",  # last-resort CLI guard must catch broad Exception
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra --strict-markers"

[tool.mypy]
python_version = "3.10"
strict = false
warn_unused_ignores = true
warn_redundant_casts = true
ignore_missing_imports = true
